+++ test/p2p/pex/dial_persistent_peers.sh
@@ -1,4 +1,4 @@
-#! /bin/bash 
+#! /bin/bash
 set -u
 
 N=$1
@@ -11,7 +11,7 @@ for i in `seq 1 $N`; do
 	curl -s $addr/status > /dev/null
 	ERR=$?
 	while [ "$ERR" != 0 ]; do
-		sleep 1	
+		sleep 1
 		curl -s $addr/status > /dev/null
 		ERR=$?
 	done
@@ -19,13 +19,13 @@ for i in `seq 1 $N`; do
 done
 
 set -e
-# seeds need quotes
-seeds="\"$(test/p2p/ip.sh 1):46656\""
+# persistent_peers need quotes
+persistent_peers="\"$(test/p2p/ip.sh 1):46656\""
 for i in `seq 2 $N`; do
-	seeds="$seeds,\"$(test/p2p/ip.sh $i):46656\""
+	persistent_peers="$persistent_peers,\"$(test/p2p/ip.sh $i):46656\""
 done
-echo $seeds
+echo $persistent_peers
 
-echo $seeds
+echo $persistent_peers
 IP=$(test/p2p/ip.sh 1)
-curl --data-urlencode "seeds=[$seeds]" "$IP:46657/dial_seeds"
+curl --data-urlencode "persistent_peers=[$persistent_peers]" "$IP:46657/dial_persistent_peers"